make sorting and genre filters as query
