+++ src/comics/mod.rs
@@ -1,13 +1,8 @@
 use axum::{http::StatusCode, response::IntoResponse};
-use chrono::{DateTime, Utc};
 use diesel::result::{DatabaseErrorKind, Error::DatabaseError};
 use diesel_async::pooled_connection::deadpool::PoolError;
-use serde::Deserialize;
-use serde::Serialize;
-use utoipa::IntoParams;
-use uuid::Uuid;
 
-use crate::{ErrorResponse, SortingOrder};
+use crate::ErrorResponse;
 
 use self::chapters::routes::FILE_SIZE_LIMIT_MB;
 
@@ -17,42 +12,6 @@ pub mod comic_genres;
 pub mod models;
 pub mod routes;
 mod utils;
-
-#[derive(Debug, Deserialize, IntoParams)]
-pub struct ComicsParams {
-    #[serde(default = "Uuid::max")]
-    pub max_id: Uuid,
-    #[serde(default)]
-    pub genre: Option<i32>,
-    #[serde(default)]
-    pub sorting: Option<SortingOrder>,
-    #[serde(default)]
-    pub order: Order,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Order {
-    Latest(DateTime<chrono::Utc>),
-    Best(f64),
-}
-
-impl Default for Order {
-    fn default() -> Self {
-        Self::Latest(Utc::now())
-    }
-}
-
-impl Default for ComicsParams {
-    fn default() -> Self {
-        ComicsParams {
-            max_id: Uuid::max(),
-            genre: Option::default(),
-            order: Order::default(),
-            sorting: Option::default(),
-        }
-    }
-}
 
 #[derive(thiserror::Error, Debug)]
 pub enum ComicsError {
